+++ docs/src/helpers/playground.js
@@ -13,12 +13,12 @@ export const hideComponentBoilerplate = (code) => {
 
   // Apply the first fold
   let foldedCode = code.replace(getTextRegex, (match) => {
-    return `// click ellipsus to show imports ${foldMarkerStart}\n\n${match.trim()}\n${foldMarkerEnd}\n\n`;
+    return `// click ellipsus to show imports ${foldMarkerStart}\n${match.trim()}\n${foldMarkerEnd}\n`;
   });
 
   // Apply the second fold
   foldedCode = foldedCode.replace(defineComponentRegex, (match) => {
-    return `// click ellipsus to show exports ${foldMarkerStart}\n\n${match}\n${foldMarkerEnd}\n`;
+    return `// click ellipsus to show exports ${foldMarkerStart}\n${match}\n${foldMarkerEnd}\n`;
   });
 
   return foldedCode;
@@ -64,23 +64,23 @@ export const getExampleFiles = async({
         return get(contentTypes, extension) || 'text/html';
       };
 
-      if(inArray(fileName, ['index.html'])) {
+      if(inArray(fileName, ['page.html'])) {
         const fileContent = await file();
-        exampleFiles['index.html'] = {
+        exampleFiles['page.html'] = {
           contentType: 'text/html',
           content: fileContent.default
         };
       }
-      else if(inArray(fileName, ['index.css'])) {
+      else if(inArray(fileName, ['page.css'])) {
         const fileContent = await file();
-        exampleFiles['index.css'] = {
+        exampleFiles['page.css'] = {
           contentType: 'text/css',
           content: fileContent.default
         };
       }
-      else if(inArray(fileName, ['index.js'])) {
+      else if(inArray(fileName, ['page.js'])) {
         const fileContent = await file();
-        exampleFiles['index.js'] = {
+        exampleFiles['page.js'] = {
           contentType: 'text/javascript',
           content: fileContent.default
         };
@@ -119,10 +119,17 @@ export const getExampleFiles = async({
           content: fileContent.default
         };
       }
+      if(includeLog && inArray(fileName, ['index.js', `${contentID}.js`])) {
+        const fileContent = await file();
+        exampleFiles['index.js'] = {
+          contentType: 'text/javascript',
+          content: fileContent.default
+        };
+      }
     }
   });
-  // auto generate index.html if not specified for component
-  if(!exampleFiles['index.html']?.content) {
+  // auto generate page.html if not specified for component
+  if(!exampleFiles['page.html']?.content) {
 
     // get tag name from contents
     let tagName;
@@ -136,7 +143,7 @@ export const getExampleFiles = async({
       }
     }
 
-    exampleFiles['index.html'] = {
+    exampleFiles['page.html'] = {
       contentType: 'text/html',
       generated: true,
       content: (tagName)
@@ -160,9 +167,9 @@ export const getExampleFiles = async({
     };
   }
 
-  // auto generate index.css/js if not specified for component
-  if(!exampleFiles['index.css']?.content) {
-    exampleFiles['index.css'] = {
+  // auto generate page.css/js if not specified for component
+  if(!exampleFiles['page.css']?.content) {
+    exampleFiles['page.css'] = {
       contentType: 'text/css',
       generated: true,
       content: ''
@@ -175,13 +182,14 @@ export const getExampleFiles = async({
       content: ''
     };
   }
-  if(!exampleFiles['index.js']?.content) {
-    exampleFiles['index.js'] = {
+  if(!exampleFiles['page.js']?.content) {
+    exampleFiles['page.js'] = {
       contentType: 'text/javascript',
       generated: true,
       content: ''
     };
   }
+
 
   if(includePlaygroundInjections) {
     addPlaygroundInjections(exampleFiles, { includeLog });
@@ -189,16 +197,22 @@ export const getExampleFiles = async({
 
   if(emptyIfAllGenerated) {
     let allGenerated = true;
-    each(exampleFiles, (file) => {
-      if(!file.generated) {
+    each(exampleFiles, (file, name) => {
+      if(file.generated !== true) {
         allGenerated = false;
+        return;
       }
     });
     if(allGenerated) {
-      exampleFiles = {};
+      return {};
     }
   }
-
+  if(!includeLog && exampleFiles['page.js'].generated && exampleFiles['page.css']) {
+    exampleFiles['page.html'].generated = false;
+  }
+  if(includeLog) {
+    exampleFiles['page.html'].hidden = true;
+  }
   return exampleFiles;
 };
 
@@ -221,15 +235,15 @@ export const getEmptyProjectFiles = ({
       contentType: 'text/css',
       content: '',
     },
-    'index.js': {
+    'page.js': {
       contentType: 'text/javascript',
       content: '',
     },
-    'index.html': {
+    'page.html': {
       contentType: 'text/html',
       content: '',
     },
-    'index.css': {
+    'page.css': {
       contentType: 'text/css',
       content: '',
     },
@@ -249,7 +263,7 @@ export const getPanelIndexes = (files = {}, { type } = {}) => {
   let indexes;
   if(type == 'log') {
     indexes = {
-      'index.js': 0,
+      'page.js': 0,
     };
   }
   else {
@@ -257,9 +271,9 @@ export const getPanelIndexes = (files = {}, { type } = {}) => {
       'component.js': 0,
       'component.html': 0,
       'component.css': 0,
-      'index.html': 1,
-      'index.css': 1,
-      'index.js': 1,
+      'page.html': 1,
+      'page.css': 1,
+      'page.js': 1,
     };
   }
   // filter out generated and absent files
@@ -270,7 +284,7 @@ export const getPanelIndexes = (files = {}, { type } = {}) => {
     return true;
   });
   // use right pane for css if no index files
-  if(!indexes['index.html'] && !indexes['index.css'] && !indexes['index.js']) {
+  if(!indexes['page.html'] && !indexes['page.css'] && !indexes['page.js']) {
     indexes['component.css'] = 1;
   }
   return indexes;