+++ RELEASE-NOTES.md
@@ -5,6 +5,12 @@ This is a pre-release version and APIs will change quickly. Before `1.0` release
 Please note after `1.0` Semver will be followed using normal protocols.
 
 
+# Version 0.5.1
+
+* Standard template syntax now defaults to reactiveData to avoid confusion when using subtemplates naively. You can still specify non reactive data using verbose syntax.
+* Fix issue with js expression passthrough to snippets and subtemplates
+* Added cyclical detection to to expression lookup to avoid recursion
+
 # Version 0.5.0
 
 ### Breaking Changes